V3.0.0 Release
Breaking Changes
I discovered a bug in the TLV serialization - I totally omitted the Message ID from the serialized representation, making it impossible to write a generic decoder. This was a big enough miss it is worth rolling a major version change and updating the wire format.

To make sure this doesn't happen in the future, I added the message ID to the Crunch header, and no longer delegate message header serialization/deserialization to the Serdes policy - it is now handled by the top-level APIs.

Features
Added a Decoder class to the public API. This class takes in a Serdes and Integrity policy and a parameter pack of Message classes as template parameters. It exposes a Decode API which takes in a buffer of data and a std::variant reference to populate with the decoded message. This interface makes it easy to handle multiple message types coming from one data source.

NOTE: This is a squash of many commits. In preparation for moving to a
PR-based workflow with squashed commits for a clean history, I'm
cleaning up the commits for the first three releases.

serialize message ID in TLV layout

centralize header operations into crunch/core/crunch_header.hpp

working decoder

decoder docs